requirements.txt
Django>=1.6
Pillow>=2.4.0
